% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{core_range_pal}
\alias{core_range_pal}
\alias{munro_pal}
\title{Functions for accessing palettes}
\usage{
core_range_pal(n)

munro_pal(n)
}
\arguments{
\item{n}{number of colours to return}
}
\description{
Functions for accessing palettes
}
