% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_fill_core_range}
\alias{scale_fill_core_range}
\alias{scale_colour_core_range}
\alias{scale_color_core_range}
\alias{scale_fill_munro}
\alias{scale_colour_munro}
\alias{scale_color_munro}
\title{Scale functions to use with GGPlot2}
\usage{
scale_fill_core_range(...)

scale_colour_core_range(...)

scale_color_core_range(...)

scale_fill_munro(...)

scale_colour_munro(...)

scale_color_munro(...)
}
\arguments{
\item{...}{see ggplot2 documentation}
}
\description{
scale_fill_* and scale_colour_* functions
}
