% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\docType{data}
\name{CoreRange}
\alias{CoreRange}
\alias{Munro}
\title{8 Degrees Palettes}
\format{
An object of class \code{character} of length 6.

An object of class \code{character} of length 8.
}
\usage{
CoreRange

Munro
}
\description{
CoreRange: 6 colour palette based on the core range from 8 degrees brewing

Munro: 8 colour paried paletted based on the limited edition Munro series
}
\keyword{datasets}
